+++ test/untilMorning.js
@@ -3,12 +3,10 @@ const Settings = require('./../app/utils/settings')
 const chai = require('chai')
 const mockSettingsFilePath = `${__dirname}/assets/settings.untilMorning.json`
 const fs = require('fs')
+const { DateTime } = require('luxon')
 
 chai.should()
 const timeout = process.env.CI ? 30000 : 10000
-
-const ONE_DAY = 24 * 60 * 60 * 1000
-const timezoneOffset = (new Date().getTimezoneOffset()) / 60
 
 describe('UntilMorning', function () {
   let settings
@@ -31,25 +29,14 @@ describe('UntilMorning', function () {
       settings = new Settings(mockSettingsFilePath)
     })
 
-    it('timeUntilMorning() calculates time until morning', function () {
-      const currentTime = Date.now()
-      let sunrise
-
-      // Checking against UTC time
-      if (Date.now() < new Date().setHours(6, 0, 0, 0)) {
-        sunrise = new Date().setHours(6, 0, 0, 0)
-      } else {
-        sunrise = new Date(currentTime + ONE_DAY).setHours(6, 0, 0, 0)
-      }
-
-      const actual = new UntilMorning(settings).timeUntilMorning()
-      const expected = sunrise - currentTime
-
-      Math.abs(expected - actual).should.be.lessThan(1000)
+    it('msToSunrise() returns morning time the same day', function () {
+      const dt = DateTime.local().set({ hours: 5, minutes: 0, seconds: 0 })
+      new UntilMorning(settings).msToSunrise(dt).should.be.within(60 * 60 * 1000 - 60000, 60 * 60 * 1000 + 60000)
     })
 
-    it('loadMorningTime() returns morning time', function () {
-      new UntilMorning(settings).loadMorningTime(new Date()).hour().should.equal(6 + timezoneOffset)
+    it('msToSunrise() returns morning time the next day', function () {
+      const dt = DateTime.local().set({ hours: 7, minutes: 0, seconds: 0 })
+      new UntilMorning(settings).msToSunrise(dt).should.be.within(23 * 60 * 60 * 1000 - 60000, 23 * 60 * 60 * 1000 + 60000)
     })
   })
 
@@ -61,28 +48,22 @@ describe('UntilMorning', function () {
       settings = new Settings(mockSettingsFilePath)
     })
 
-    // Checking against UTC time
-    it('timeUntilMorning() calculates time until morning when input is a random hour', function () {
-      let sunrise
-
-      if (Date.now() < new Date().setHours(15, 0, 0, 0)) {
-        sunrise = new Date().setHours(15, 0, 0, 0)
-      } else {
-        sunrise = new Date(Date.now() + ONE_DAY).setHours(15, 0, 0, 0)
-      }
-
-      const expected = sunrise - new Date()
-      const actual = new UntilMorning(settings).timeUntilMorning()
-
-      Math.abs(expected - actual).should.be.lessThan(1000)
+    it('msToSunrise() returns morning time the same day', function () {
+      const dt = DateTime.local().set({ hours: 14, minutes: 0, seconds: 0 })
+      new UntilMorning(settings).msToSunrise(dt).should.be.within(60 * 60 * 1000 - 60000, 60 * 60 * 1000 + 60000)
     })
 
-    it('loadMorningTime() returns morning time', function () {
-      new UntilMorning(settings).loadMorningTime(new Date()).hour().should.equal(15 + timezoneOffset)
+    it('msToSunrise() returns morning time the next day', function () {
+      const dt = DateTime.local().set({ hours: 16, minutes: 0, seconds: 0 })
+      new UntilMorning(settings).msToSunrise(dt).should.be.within(23 * 60 * 60 * 1000 - 60000, 23 * 60 * 60 * 1000 + 60000)
     })
   })
 
   describe('Sunrise Settings', function () {
+    // test for when the functionality to input sunrise is added
+    // test data for 08/04/2018 Amsterdam, Netherlands
+    // expected time is August 4 2018, 06:10 local Amsterdam time
+
     beforeEach(() => {
       createSettingsFile(mockSettingsFilePath, {
         morningHour: 'sunrise',
@@ -92,18 +73,14 @@ describe('UntilMorning', function () {
       settings = new Settings(mockSettingsFilePath)
     })
 
-    it('loadMorningTime() returns morning time when sunrise is an input', function () {
-      // test for when the functionality to input sunrise is added
-      // test data for 08/04/2018 Amsterdam, Netherlands
-      // Checking against UTC time: expected time is August 4 2018, 04:05:0000 UTC
-      const date = new Date()
-      date.setFullYear(2018)
-      date.setDate(4)
-      date.setMonth(7)
+    it('msToSunrise() returns morning time the same day', function () {
+      const dt = DateTime.local(2018, 8, 4, 5, 10, 0, 0)
+      new UntilMorning(settings).msToSunrise(dt).should.be.within(60 * 60 * 1000 - 60000, 60 * 60 * 1000 + 60000)
+    })
 
-      const morning = new UntilMorning(settings).loadMorningTime(date)
-      morning.hour().should.equal(4)
-      morning.minute().should.equal(9)
+    it('msToSunrise() returns morning time the next day', function () {
+      const dt = DateTime.local(2018, 8, 4, 7, 10, 0, 0)
+      new UntilMorning(settings).msToSunrise(dt).should.be.within(23 * 60 * 60 * 1000 - 60000, 23 * 60 * 60 * 1000 + 60000)
     })
   })
 })